typecast spaces to print to int_8, strlen > 5 would cause a bug with the old implementation pidMax was printed as temperature instead of temperature difference

+++ TemperatureFormats.cpp
@@ -118,7 +118,7 @@ fixed23_9 stringToFixedPoint(const char * numberString){
 	if(fractPtr != 0){
 		// decimal point was found
 		fractPtr++; // add 1 to pointer to skip point
-		uint8_t numDecimals = strlen(fractPtr);
+		int8_t numDecimals = (int8_t) strlen(fractPtr);
 		fracPart = atol(fractPtr);		
 		fracPart = fracPart << 9; // 9 bits for fraction part
 		while(numDecimals > 0){
